+++ include/Bulwark.h
@@ -3,8 +3,8 @@
 
 #include "Color16Definitions.h"
 
-#include <lang/types.h>
 #include <stdio.h>
+#include <stdbool.h>
 
 typedef struct BulwarkEvent BulwarkEvent;
 
@@ -24,15 +24,15 @@ BulwarkEventType BulwarkEvent_GetType(BulwarkEvent *event);
 char BulwarkEvent_GetCharacter(BulwarkEvent *event);
 
 /* Color functions */
-void Bulwark_SetForegroundColor16(uint8 color16);
-void Bulwark_SetBackgroundColor16(uint8 color16);
-void Bulwark_SetForegroundAndBackgroundColors16(uint8 foregroundColor16, uint8 backgroundColor16);
+void Bulwark_SetForegroundColor16(int color16);
+void Bulwark_SetBackgroundColor16(int color16);
+void Bulwark_SetForegroundAndBackgroundColors16(int foregroundColor16, int backgroundColor16);
 void Bulwark_ClearForegroundColor();
 void Bulwark_ClearBackgroundColor();
 void Bulwark_ClearForegroundAndBackgroundColor();
 
 /* Drawing functions */
-void Bulwark_SetDrawPosition(uint16 x, uint16 y);
+void Bulwark_SetDrawPosition(int x, int y);
 void Bulwark_DrawCharacter(char character);
 void Bulwark_DrawString(const char *string);
 
@@ -40,8 +40,8 @@ void Bulwark_DrawString(const char *string);
 void Bulwark_SetCursorVisible(bool cursorVisible);
 
 /* Window functions */
-uint16 Bulwark_GetWindowWidth();
-uint16 Bulwark_GetWindowHeight();
+int Bulwark_GetWindowWidth();
+int Bulwark_GetWindowHeight();
 
 /* Sound functions */
 void Bulwark_PlayBellSound();
